Neaten code and resolve fatal error.
Fatal error was the _audio.PlayPredicted line trying to play the pickup sound at the location of the deck. When you take the last card that deletes the deck and it tried to find the coordinates of a entity that doesn't exist. Causing fatal crash. I changed the location from the card deck to the player. This should solve the issue and since the player has to be near the deck to interact with it, it shouldn't be issue.

Can't abstract some of these entities. Thought I could, but it doesn't like it.

I also merged some lines that I'm pretty sure could be merged. Let me know if I fucked up cause I can't always tell.