Fix Nuno's Sprite Layering, Tweak Nuno Sprites
Changed the way the layering is implemented in yaml so it actually works.
Made the playing cards 1 pixel taller. Why? Don't ask.
Modified all numbers for clarity and to be centered in the card.
Created a new Reverse sprite because the old one kinda sucks.
Changed the block card to skip. (It is called skip in the yaml).
Remove the border sprite and merge it into the background.